requirements.txt
absl-py==0.7.1
astor==0.8.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bz2file==0.98
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Cython==0.29.12
dill==0.3.0
enum34==1.1.6
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.22.0
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.4
mock==3.0.5
numpy==1.16.4
object-detection==0.1
Pillow==6.1.0
promise==2.2.1
protobuf==3.8.0
psutil==5.6.3
pyparsing==2.4.0
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
subprocess32==3.5.4
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==2.0.0b1
tensorflow-datasets==1.0.2
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-metadata==0.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tqdm==4.32.2
typing==3.7.4
urllib3==1.25.3
Werkzeug==0.15.4
wget==3.2
wrapt==1.11.2
xtract==0.1a3
